flutter_wth_django/django/api_project/requirements.txt
# --- to run this enter the command in your terminal  
# =>  pip install -r requirements.txt

# Django and Django REST Framework
django==5.1.3
djangorestframework==3.15.2
